Lag behind realtime by (cfgable) interval
We noticed that in some cases there would be gaps in the log entries
read by the reader. You can observe this same behavior by using the web
based stackdriver UI and repeatedly refreshing for new logs.

To address this, add a configurable lag interval. When reading log
entries, only read up to now-the lag interval. The default is 30
seconds, meaning that the bridge will always read entries up to 30
seconds ago.