Add better config management
Add a better config object with config read from a config file and/or
command line args, using viper.

Command line args override options in config file which in turn override
defaults.

Add a configmap that gets deployed alongside the bridge which allows
for easier configuration of the bridge. The configmap is optional so if
not specified the defaults will be used.

Command line options continue to work so they could also be specified in
the deployment yaml if desired.