Handle errs!=iterator.Done(e.g. ResourceExhausted)
Properly handle errors that are not iterator.Done when iterating over
entries via it.Next().

If the error is iterator.Done, break the loop like before. Otherwise, if
the error is non-nil, log a message and break the loop. To reduce log
file pollution, the number of sequential errors are counted. The log
level is debug for the first retry attempt and warn afterward. The
counter is reset every time it.Next() returns a log entry.

This fixes
https://github.com/sysdiglabs/stackdriver-webhook-bridge/issues/11.